requirements.txt
torch==2.1.2
torchvision==0.16.2
timm==0.9.10
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.1
tqdm==4.66.1
pillow==10.0.0
rasterio==1.3.8
urllib3==2.0.4
requests==2.31.0
tensorboard==2.14.0
einops==0.7.0
